requirements.txt
flask
flask-cors
openai
langchain
tiktoken
supabase
chromadb
python-dotenv